video-7 (Limitless Limits for Eager Loading)
Limiting the number of records on an eager loaded relationship in Laravel has historically been a pain, with a number of unexpected side effects. In Laravel 11, we can finally lay that issue to rest, because it now works exactly as you’d expect.

return User::with(['posts' => fn($query) => $query->latest()->limit(2)])->get();

here we apply array syntax for posts so we can apply custom logic to it.
in previous version it just load the logic for the first index but not for others because og limits in SQL. you have to install severl pacakges for it.
in laravel 11 the problem is solved now it changes it work directly.

we can also achive this on relation in model.
public function latestPosts(): HasMany
{
    return $this->hasMany(Post::class)->latest()->limit(2);
}