Video 14 (Simple Tests for Complex Jobs)
Have you ever wanted to test that a job has been marked as deleted, failed or released? Doing so has been a bit of hack in previous Laravel versions. A shiny withFakeQueueInteractions method makes that frustration a thing of the past. For our final episode, let's look at an example.